переделал используя анонимные классы
