+++ src/ru/avalon/vergentev/j110/labwork5_1/Main.java
@@ -3,30 +3,55 @@ package ru.avalon.vergentev.j110.labwork5_1;
 public class Main {
     public static void main(String[] args) {
 
-        UsualClass y1 = new FirstEquation(0.01, 2, 0.00001);
-        UsualClass y2 = new SecondEquation(2, 3, 0.00001);
-        UsualClass y3 = new ThirdEquation(1, 3, 0.00001);
-        UsualClass y4 = new FourthEquation(2, 4, 0.00001);
-        UsualClass y5 = new FifthEquation(1, 5, 0.00001);
+        //формируем анонимные классы:
+        UsualClass equation1 = new UsualClass(0.01, 2, 0.00001) {
+            @Override
+            public double y(double x) {
+                return (Math.exp(-x) - 0.5);
+            }
+        };
+        UsualClass equation2 = new UsualClass(2, 3, 0.00001) {
+            @Override
+            public double y(double x) {
+                return (Math.sin(x)-0.75);
+            }
+        };
+        UsualClass equation3 = new UsualClass(1, 3, 0.00001) {
+            @Override
+            public double y(double x) {
+                return (Math.log(x*x*x)-2);
+            }
+        };
+        UsualClass equation4 = new UsualClass(2, 4, 0.00001) {
+            @Override
+            public double y(double x) {
+                return (Math.tan(x));
+            }
+        };
+        UsualClass equation5 = new UsualClass(1, 5, 0.00001) {
+            @Override
+            public double y(double x) {
+                return (x*x*x - 8*x + 2);
+            }
+        };
 
-        //вызов из классов наследников (ссылка на метод экземпл€ра)
-        System.out.println("From Extends classes:");
-        y1.halfDivision();
-        y2.halfDivision();
-        y3.halfDivision();
-        y4.halfDivision();
-        y5.halfDivision();
+        //вызываем методы анонимных классов ссыла€сь на экземпл€р класса
+        System.out.println("Instance method reference: ");
+        equation1.halfDivision();
+        equation2.halfDivision();
+        equation3.halfDivision();
+        equation4.halfDivision();
+        equation5.halfDivision();
         System.out.print('\n');
 
-        //вызов статического метода
-        System.out.println("From Extends classes with static method:");
-        UsualClass.halfDivisionStatic(y1);
-        UsualClass.halfDivisionStatic(y2);
-        UsualClass.halfDivisionStatic(y3);
-        UsualClass.halfDivisionStatic(y4);
-        UsualClass.halfDivisionStatic(y5);
+        //вызываем методы анонимных классов через статический метод
+        System.out.println("Static method reference: ");
+        UsualClass.halfDivisionStatic(equation1);
+        UsualClass.halfDivisionStatic(equation2);
+        UsualClass.halfDivisionStatic(equation3);
+        UsualClass.halfDivisionStatic(equation4);
+        UsualClass.halfDivisionStatic(equation5);
         System.out.print('\n');
-
 
     }
 }
